Rename cqc locations cleaned column names file
