+++ jobs/flatten_cqc_ratings.py
@@ -13,7 +13,7 @@ from utils import (
 from utils.column_names.raw_data_files.cqc_location_api_columns import (
     NewCqcLocationApiColumns as CQCL,
 )
-from utils.column_names.cleaned_data_files.cqc_location_cleaned_values import (
+from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
     CqcLocationCleanedValues as CQCLValues,
 )
 from utils.column_names.raw_data_files.ascwds_workplace_columns import (
