+++ jobs/merge_ind_cqc_data.py
@@ -6,7 +6,7 @@ from pyspark.sql.types import FloatType, StringType
 from utils import utils
 import utils.cleaning_utils as cUtils
 
-from utils.column_names.cleaned_data_files.cqc_location_cleaned_values import (
+from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
     CqcLocationCleanedColumns as CQCLClean,
     CqcLocationCleanedValues as CQCLValues,
 )
