+++ tests/test_file_schemas.py
@@ -56,7 +56,7 @@ from utils.column_names.cleaned_data_files.cqc_pir_cleaned_values import (
 from utils.column_names.cleaned_data_files.cqc_provider_cleaned_values import (
     CqcProviderCleanedColumns as CQCPClean,
 )
-from utils.column_names.cleaned_data_files.cqc_location_cleaned_values import (
+from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
     CqcLocationCleanedColumns as CQCLClean,
 )
 from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
