+++ tests/unit/test_reconciliation.py
@@ -14,7 +14,7 @@ from utils.column_names.raw_data_files.cqc_location_api_columns import (
 from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
     AscwdsWorkplaceCleanedColumns as AWPClean,
 )
-from utils.column_names.cleaned_data_files.cqc_location_cleaned_values import (
+from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
     CqcLocationCleanedColumns as CQCLClean,
 )
 from utils.reconciliation_utils.reconciliation_values import (
