+++ utils/validation/validation_rules/locations_api_cleaned_validation_rules.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from utils.column_names.cleaned_data_files.cqc_location_cleaned_values import (
+from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
     CqcLocationCleanedColumns as CQCLClean,
 )
 from utils.column_values.categorical_columns_by_dataset import (
